+++ .changeset/pink-suits-flow.md
@@ -1,0 +1,5 @@
+---
+"@ethlete/types": patch
+---
+
+Update to latest ethlete types
