+++ libs/types/src/lib/api/Overall/statisticOverallUtCardResults.view.ts
@@ -1,5 +1,5 @@
-// Generated Interface, do not change manually!
-import { MediaMinimalView } from './mediaMinimal.view';
+import { MediaMinimalView } from "../Media";
+
 
 export interface StatisticOverallUtCardResultsView {
     goalsShot: number | null;
