+++ libs/types/src/lib/api/Round/roundWithStageAndGroup.view.ts
@@ -1,6 +1,7 @@
 // Generated Interface, do not change manually!
-import { StageListView } from './stageList.view';
-import { GroupListView } from './groupList.view';
+
+import GroupListView from "../Group/groupList.view";
+import StageListView from "../Stage/stageList.view";
 
 export interface RoundWithStageAndGroupView {
     id: string;
