+++ libs/types/src/lib/api/Tournament/Registration/userTournamentRegistration.view.ts
@@ -1,6 +1,7 @@
 // Generated Interface, do not change manually!
+import { MediaView } from '../../Media';
 import { CheckInStatus } from './../../Registration/checkIn.status';
-import { MediaView } from './media.view';
+
 
 export interface UserTournamentRegistrationView {
     id: string;
