+++ libs/types/src/lib/api/Weekly/statisticWeeklyUtCardResults.view.ts
@@ -1,5 +1,7 @@
 // Generated Interface, do not change manually!
-import { MediaMinimalView } from './mediaMinimal.view';
+
+import { MediaMinimalView } from "../Media";
+
 
 export interface StatisticWeeklyUtCardResultsView {
     goalsShot: number | null;
